requirements.txt
Django==1.6
Flask==0.10.1
Flask-Admin==1.0.7
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-RESTful==0.2.6
Flask-SQLAlchemy==1.0
Flask-Script==0.6.3
Flask-WTF==0.9.3
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
MySQL-python==1.2.4
PyYAML==3.10
SQLAlchemy==0.8.3
South==0.8.3
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.0
amqp==1.3.3
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.7
blinker==1.3
boto==2.17.0
celery==3.1.4
django-celery==3.1.1
django-grappelli==2.5.0
django-kombu==0.9.4
django-storages==1.1.8
django-timezone-field==1.0
django-widget-tweaks==1.3
feedparser==5.1.3
httplib2==0.8
ipython==1.1.0
itsdangerous==0.23
kombu==3.0.5
librabbitmq==1.0.3
mock==1.0.1
nltk==2.0.4
nose==1.3.0
numpy==1.8.0
oauth2==1.5.211
oauthlib==0.6.0
py-bcrypt==0.4
pygeoip==0.3.0
pyjade==2.0.3
python-dateutil==2.2
pytz==2013.8
rauth==0.6.2
readline==6.2.4.1
requests==2.0.1
requests-cache==0.4.4
requests-oauthlib==0.4.0
six==1.4.1
twitter-text-python==1.0.1
twython==3.1.0
uWSGI==1.9.20
wsgiref==0.1.2
wtf==0.1
